refactor(db): move AuditLogDetails type to the schema file
`schema/auditLogs.ts` was importing `AuditLogDetails` from
`utils/audit.ts`, which makes the schema layer depend on utils
(the wrong direction — utils should depend on schema, not the
reverse). Move the type to live alongside the table definition
that uses it, and have `utils/audit.ts` and the repo import it
from the schema instead.
